fix(sea): rebind llm helper symbols in runtime wrapper for stale environments
