+++ sea/runtime.py
@@ -1,23 +1,31 @@
 from __future__ import annotations
 
+import asyncio
+import json
 import logging
 import os
-import uuid
-import asyncio
-from datetime import datetime, timezone as dt_timezone
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Callable
-import json
 import re
+import uuid
+from datetime import datetime
+from datetime import timezone as dt_timezone
+from pathlib import Path
+from typing import Any, Callable, Dict, List, Optional
 
+from database.models import Playbook as PlaybookModel
 from llm_clients.exceptions import LLMError
 from saiverse.logging_config import log_sea_trace
-from sea.playbook_models import NodeType, PlaybookSchema, PlaybookValidationError, validate_playbook_graph
-from sea.langgraph_runner import compile_playbook
-from sea.cancellation import CancellationToken, ExecutionCancelledException
-from database.models import Playbook as PlaybookModel
 from saiverse.model_configs import get_model_parameter_defaults
 from saiverse.usage_tracker import get_usage_tracker
+from sea.cancellation import CancellationToken, ExecutionCancelledException
+from sea.langgraph_runner import compile_playbook
+from sea.playbook_models import NodeType, PlaybookSchema, PlaybookValidationError, validate_playbook_graph
+from sea.runtime_context import prepare_context as prepare_context_impl
+from sea.runtime_context import preview_context as preview_context_impl
+from sea.runtime_graph import compile_with_langgraph as compile_with_langgraph_impl
+from sea.runtime_llm import lg_llm_node as lg_llm_node_impl
+
+from . import runtime_llm as runtime_llm_module
+from .runtime_utils import _format, _is_llm_streaming_enabled
 
 LOGGER = logging.getLogger(__name__)
 
@@ -26,41 +34,6 @@ def _get_default_lightweight_model() -> str:
     """Get the default lightweight model from environment or fallback."""
     return os.getenv("SAIVERSE_DEFAULT_LIGHTWEIGHT_MODEL", "gemini-2.5-flash-lite")
 
-
-def _is_llm_streaming_enabled() -> bool:
-    """Check if LLM streaming is enabled (default: True)."""
-    val = os.getenv("SAIVERSE_LLM_STREAMING", "true")
-    result = val.lower() not in ("false", "0", "off", "no")
-    logging.info("[DEBUG] _is_llm_streaming_enabled: raw_val=%r, result=%s", val, result)
-    return result
-
-
-def _format(template: str, variables: Dict[str, Any]) -> str:
-    """Format template with variables, supporting dot notation keys.
-
-    Uses regex-based replacement to safely handle templates where variable values
-    may contain curly braces (e.g., LLM-generated text with {}).
-    """
-    result = template
-
-    # Build a lookup dict with all keys (including nested access via dot notation)
-    lookup: Dict[str, str] = {}
-    for key, value in variables.items():
-        lookup[str(key)] = str(value) if value is not None else ""
-
-    # Replace {key} patterns with corresponding values
-    # Only replace if the key exists in our lookup
-    def replacer(match: re.Match) -> str:
-        key = match.group(1)
-        if key in lookup:
-            return lookup[key]
-        # Key not found, leave placeholder as-is
-        return match.group(0)
-
-    # Pattern: {word_chars_and_dots} but not empty
-    result = re.sub(r"\{([\w.]+)\}", replacer, result)
-
-    return result
 
 
 class SEARuntime:
@@ -279,808 +252,27 @@ class SEARuntime:
         cancellation_token: Optional[CancellationToken] = None,
         pulse_type: Optional[str] = None,
     ) -> Optional[List[str]]:
-        _lg_outputs: List[str] = []
-        temperature = self._default_temperature(persona)
-        parent = parent_state or {}
-
-        # Update execution state: playbook started (LangGraph path)
-        if hasattr(persona, "execution_state"):
-            persona.execution_state["playbook"] = playbook.name
-            persona.execution_state["node"] = playbook.start_node
-            persona.execution_state["status"] = "running"
-
-        compiled = compile_playbook(
+        return compile_with_langgraph_impl(
+            self,
             playbook,
-            llm_node_factory=lambda node_def: self._lg_llm_node(node_def, persona, building_id, playbook, event_callback),
-            tool_node_factory=lambda node_def: self._lg_tool_node(node_def, persona, playbook, event_callback),
-            tool_call_node_factory=lambda node_def: self._lg_tool_call_node(node_def, persona, playbook, event_callback),
-            speak_node=lambda state: self._lg_speak_node(state, persona, building_id, playbook, _lg_outputs, event_callback),
-            think_node=lambda state: self._lg_think_node(state, persona, playbook, _lg_outputs, event_callback),
-            say_node_factory=lambda node_def: self._lg_say_node(node_def, persona, building_id, playbook, _lg_outputs, event_callback),
-            memorize_node_factory=lambda node_def: self._lg_memorize_node(node_def, persona, playbook, _lg_outputs, event_callback),
-            exec_node_factory=lambda node_def: self._lg_exec_node(node_def, playbook, persona, building_id, auto_mode, _lg_outputs, event_callback),
-            subplay_node_factory=lambda node_def: self._lg_subplay_node(node_def, persona, building_id, playbook, auto_mode, _lg_outputs, event_callback),
-            set_node_factory=lambda node_def: self._lg_set_node(node_def, playbook, event_callback),
-            stelis_start_node_factory=lambda node_def: self._lg_stelis_start_node(node_def, persona, playbook, event_callback),
-            stelis_end_node_factory=lambda node_def: self._lg_stelis_end_node(node_def, persona, playbook, event_callback),
+            persona,
+            building_id,
+            user_input,
+            auto_mode,
+            base_messages,
+            pulse_id,
+            parent_state=parent_state,
+            event_callback=event_callback,
+            cancellation_token=cancellation_token,
+            pulse_type=pulse_type,
         )
-        if not compiled:
-            # Update execution state: compilation failed, reset to idle
-            if hasattr(persona, "execution_state"):
-                persona.execution_state["playbook"] = None
-                persona.execution_state["node"] = None
-                persona.execution_state["status"] = "idle"
-            raise LLMError(
-                f"Playbook '{playbook.name}' graph compilation failed",
-                user_message=f"プレイブック '{playbook.name}' のグラフ構築に失敗しました。",
-            )
-
-        # Process input_schema to inherit variables from parent_state
-        inherited_vars = {}
-        for param in playbook.input_schema:
-            param_name = param.name
-            source_key = param.source if param.source else "input"
-
-            # If this param already exists in parent (e.g., from initial_params/playbook_params),
-            # use that value instead of resolving from source
-            if param_name in parent and parent[param_name] is not None:
-                value = parent[param_name]
-                LOGGER.debug("[sea][LangGraph] Using existing value for %s from parent: %s", param_name, str(value) if value else "(empty)")
-            # Resolve value from parent_state or fallback
-            elif source_key.startswith("parent."):
-                actual_key = source_key[7:]  # strip "parent."
-                # Use _resolve_state_value for dot-notation support (e.g., "current_task.objective")
-                value = self._resolve_state_value(parent, actual_key)
-                if value is None:
-                    value = ""
-                LOGGER.debug("[sea][LangGraph] Resolved %s from parent.%s: %s", param_name, actual_key, str(value) if value else "(empty)")
-            elif source_key == "input":
-                value = user_input or ""
-            else:
-                value = parent.get(source_key, "")
-
-            inherited_vars[param_name] = value
-
-        # Inherit pulse_usage_accumulator from parent_state if it exists (for sub-playbook calls)
-        # This ensures usage is accumulated across all LLM calls in the entire pulse chain
-        parent_accumulator = parent.get("pulse_usage_accumulator")
-        if parent_accumulator:
-            # Use the same accumulator (reference) to accumulate across sub-playbooks
-            usage_accumulator = parent_accumulator
-        else:
-            # Create new accumulator for this pulse
-            usage_accumulator = {
-                "total_input_tokens": 0,
-                "total_output_tokens": 0,
-                "total_cached_tokens": 0,
-                "total_cache_write_tokens": 0,
-                "total_cost_usd": 0.0,
-                "call_count": 0,
-                "models_used": [],
-            }
-
-        # Inherit activity trace list (shared reference, same pattern as accumulator)
-        parent_activity_trace = parent.get("_activity_trace")
-        if parent_activity_trace is not None:
-            activity_trace = parent_activity_trace
-        else:
-            activity_trace = []
-
-        initial_state = {
-            "messages": list(base_messages),
-            "inputs": {"input": user_input or ""},
-            "context": {},
-            "last": user_input or "",
-            "outputs": _lg_outputs,
-            "persona_obj": persona,
-            "pulse_id": pulse_id,
-            "pulse_type": pulse_type,  # user/schedule/auto
-            "_cancellation_token": cancellation_token,  # For node-level cancellation checks
-            "pulse_usage_accumulator": usage_accumulator,  # Inherit from parent or create new
-            "_activity_trace": activity_trace,  # Shared trace of exec/tool activities
-            "_intermediate_msgs": [],  # Track intermediate node outputs for profile-based context
-            **inherited_vars,  # Add inherited variables from input_schema
-        }
-
-        # Execute compiled playbook
-        # Set recursion limit high enough for agentic loops (default is 25, too low for multi-step agents)
-        langgraph_config = {"recursion_limit": 1000}
-
-        try:
-            # Check cancellation before starting execution
-            if cancellation_token:
-                cancellation_token.raise_if_cancelled()
-            
-            # Check if we're inside an existing event loop
-            try:
-                running_loop = asyncio.get_running_loop()
-            except RuntimeError:
-                running_loop = None
-
-            if running_loop and running_loop.is_running():
-                # We're inside an existing loop (e.g., Gradio), use run_in_executor
-                import concurrent.futures
-                with concurrent.futures.ThreadPoolExecutor() as executor:
-                    future = executor.submit(asyncio.run, compiled(initial_state, langgraph_config))
-                    final_state = future.result()
-            else:
-                # No running loop, use asyncio.run directly
-                final_state = asyncio.run(compiled(initial_state, langgraph_config))
-        except ExecutionCancelledException:
-            # Re-raise cancellation exceptions
-            raise
-        except LLMError:
-            # Re-raise LLM errors for proper error handling in caller
-            # Update execution state: execution failed, reset to idle
-            if hasattr(persona, "execution_state"):
-                persona.execution_state["playbook"] = None
-                persona.execution_state["node"] = None
-                persona.execution_state["status"] = "idle"
-            raise
-        except Exception as exc:
-            LOGGER.exception("SEA LangGraph execution failed")
-            # Update execution state: execution failed, reset to idle
-            if hasattr(persona, "execution_state"):
-                persona.execution_state["playbook"] = None
-                persona.execution_state["node"] = None
-                persona.execution_state["status"] = "idle"
-            # Wrap as LLMError so existing error propagation chain
-            # delivers it to the frontend instead of silently swallowing.
-            raise LLMError(
-                f"Playbook execution failed: {type(exc).__name__}: {exc}",
-                original_error=exc,
-                user_message=f"プレイブックの実行中にエラーが発生しました: {exc}",
-            ) from exc
-
-        # Write back state variables to parent_state based on output_schema
-        if parent_state is not None and isinstance(final_state, dict) and playbook.output_schema:
-            for key in playbook.output_schema:
-                if key in final_state:
-                    value = final_state[key]
-                    # Use _store_structured_result to also create flattened dot-notation keys
-                    # (e.g., research_result.summary, research_result.status)
-                    if isinstance(value, dict):
-                        self._store_structured_result(parent_state, key, value)
-                    else:
-                        parent_state[key] = value
-                    LOGGER.debug("[sea][LangGraph] Propagated %s to parent_state: %s", key, str(value))
-
-        # Update execution state: playbook completed (LangGraph path)
-        if hasattr(persona, "execution_state"):
-            persona.execution_state["playbook"] = None
-            persona.execution_state["node"] = None
-            persona.execution_state["status"] = "idle"
-
-        # speak/think nodes already emitted; return collected texts for UI consistency
-        return list(_lg_outputs)
 
     def _lg_llm_node(self, node_def: Any, persona: Any, building_id: str, playbook: PlaybookSchema, event_callback: Optional[Callable[[Dict[str, Any]], None]] = None):
-        async def node(state: dict):
-            # Check for cancellation at start of node
-            cancellation_token = state.get("_cancellation_token")
-            if cancellation_token:
-                cancellation_token.raise_if_cancelled()
-            
-            # Send status event for node execution
-            node_id = getattr(node_def, "id", "llm")
-            if event_callback:
-                event_callback({"type": "status", "content": f"{playbook.name} / {node_id}", "playbook": playbook.name, "node": node_id})
-            # Merge state into variables for template formatting
-            if playbook.name == 'sub_router_user':
-                action_dbg = getattr(node_def, 'action', None)
-                LOGGER.debug('[sea][router-debug] action=%s model_type=%s avail_len=%s',
-                             (action_dbg[:120] + '...') if isinstance(action_dbg, str) and len(action_dbg) > 120 else action_dbg,
-                             getattr(node_def, 'model_type', None),
-                             len(str(state.get('available_playbooks'))) if state.get('available_playbooks') is not None else None)
-
-            # Build variables for template formatting
-            variables = {
-                "input": state.get("inputs", {}).get("input", ""),
-                "last": state.get("last", ""),
-                "persona_id": getattr(persona, "persona_id", None),
-                "persona_name": getattr(persona, "persona_name", None),
-                **{k: v for k, v in state.items() if k not in ["messages", "inputs", "outputs", "persona_obj", "_cancellation_token"]},  # Include all other state variables
-            }
-
-            # Debug: log template variables for novel_writing playbook
-            if playbook.name == "novel_writing":
-                node_id = getattr(node_def, "id", "")
-                if node_id.startswith("chapter_"):
-                    # Log specific variables used in chapter templates
-                    relevant_keys = ["novel_title", "chapter_1_title", "chapter_2_title", "chapter_3_title", "chapter_4_title"]
-                    relevant_vars = {k: variables.get(k) for k in relevant_keys}
-                    LOGGER.debug("[sea][novel_writing] Node %s: relevant variables = %s", node_id, relevant_vars)
-            text = ""
-            schema_consumed = False
-            prompt = None  # Will store the expanded prompt for memorize
-            try:
-                # Determine base messages: use context_profile if set, otherwise legacy state["messages"]
-                _profile_name = getattr(node_def, "context_profile", None)
-                if _profile_name:
-                    from sea.playbook_models import CONTEXT_PROFILES
-                    _profile = CONTEXT_PROFILES.get(_profile_name)
-                    if _profile:
-                        _cache_key = f"_ctx_profile_{_profile_name}"
-                        if _cache_key not in state:
-                            state[_cache_key] = self._prepare_context(
-                                persona, building_id,
-                                state.get("inputs", {}).get("input") or None,
-                                _profile["requirements"],
-                                pulse_id=state.get("pulse_id"),
-                            )
-                            LOGGER.info("[sea] Prepared context for profile '%s' (node=%s, %d messages)",
-                                        _profile_name, node_id, len(state[_cache_key]))
-                        _profile_base = state[_cache_key]
-                        _intermediate = state.get("_intermediate_msgs", [])
-                        base_msgs = list(_profile_base) + list(_intermediate)
-                    else:
-                        LOGGER.warning("[sea] Unknown context_profile '%s' on node '%s', falling back to state messages", _profile_name, node_id)
-                        base_msgs = state.get("messages", [])
-                else:
-                    base_msgs = state.get("messages", [])
-                action_template = getattr(node_def, "action", None)
-                if action_template:
-                    prompt = _format(action_template, variables)
-                    # Auto-wrap in <system> tags to distinguish from user messages
-                    if not prompt.lstrip().startswith("<system>"):
-                        prompt = f"<system>{prompt}</system>"
-                    messages = list(base_msgs) + [{"role": "user", "content": prompt}]
-                else:
-                    messages = list(base_msgs)
-
-                # Dynamically add enum to response_schema if available_playbooks exists
-                response_schema = getattr(node_def, "response_schema", None)
-                if response_schema and "available_playbooks" in state:
-                    response_schema = self._add_playbook_enum(response_schema, state.get("available_playbooks"))
-
-                # Select LLM client based on model_type and structured output needs
-                needs_structured_output = response_schema is not None
-                llm_client = self._select_llm_client(node_def, persona, needs_structured_output=needs_structured_output)
-
-                # Inject model-specific system prompt if configured
-                _model_config_key = getattr(llm_client, "config_key", None)
-                if _model_config_key:
-                    from saiverse.model_configs import get_model_system_prompt
-                    _model_sys_prompt = get_model_system_prompt(_model_config_key)
-                    if _model_sys_prompt:
-                        _injected = False
-                        for _mi, _msg in enumerate(messages):
-                            if _msg.get("role") == "system":
-                                # Create new dict to avoid mutating shared base_msgs
-                                messages[_mi] = {**_msg, "content": _msg["content"] + "\n\n---\n\n" + _model_sys_prompt}
-                                _injected = True
-                                break
-                        if not _injected:
-                            messages.insert(0, {"role": "system", "content": _model_sys_prompt})
-                        LOGGER.debug("[sea] Injected model-specific system prompt for %s", _model_config_key)
-
-                # Check if tools are available for this node
-                available_tools = getattr(node_def, "available_tools", None)
-                LOGGER.info("[DEBUG] available_tools = %s", available_tools)
-                
-                if available_tools:
-                    LOGGER.info("[DEBUG] Entering tools mode (generate with tools)")
-                    # Tool calling mode - use unified generate() with tools
-                    tools_spec = self._build_tools_spec(available_tools, llm_client)
-                    result = llm_client.generate(
-                        messages,
-                        tools=tools_spec,
-                        temperature=self._default_temperature(persona),
-                        **self._get_cache_kwargs(),
-                    )
-
-                    # Consume reasoning (thinking) from tool-mode LLM call
-                    _tool_reasoning = llm_client.consume_reasoning()
-                    _tool_reasoning_text = "\n\n".join(
-                        e.get("text", "") for e in _tool_reasoning if e.get("text")
-                    ) if _tool_reasoning else ""
-                    if _tool_reasoning_text:
-                        state["_reasoning_text"] = _tool_reasoning_text
-
-                    # Record usage
-                    usage = llm_client.consume_usage()
-                    if usage:
-                        get_usage_tracker().record_usage(
-                            model_id=usage.model,
-                            input_tokens=usage.input_tokens,
-                            output_tokens=usage.output_tokens,
-                            cached_tokens=usage.cached_tokens,
-                            cache_write_tokens=usage.cache_write_tokens,
-                            cache_ttl=usage.cache_ttl,
-                            persona_id=getattr(persona, "persona_id", None),
-                            building_id=building_id,
-                            node_type="llm_tool",
-                            playbook_name=playbook.name,
-                            category="persona_speak",
-                        )
-                        # Accumulate into pulse total
-                        from saiverse.model_configs import calculate_cost
-                        cost = calculate_cost(usage.model, usage.input_tokens, usage.output_tokens, usage.cached_tokens, usage.cache_write_tokens, cache_ttl=usage.cache_ttl)
-                        self._accumulate_usage(state, usage.model, usage.input_tokens, usage.output_tokens, cost, usage.cached_tokens, usage.cache_write_tokens)
-
-                    # Parse output_keys to determine where to store results
-                    output_keys_spec = getattr(node_def, "output_keys", None)
-                    text_key = None
-                    function_call_key = None
-                    thought_key = None
-
-                    if output_keys_spec:
-                        for mapping in output_keys_spec:
-                            if "text" in mapping:
-                                text_key = mapping["text"]
-                            if "function_call" in mapping:
-                                function_call_key = mapping["function_call"]
-                            if "thought" in mapping:
-                                thought_key = mapping["thought"]
-
-                    # Debug: log result type and keys
-                    LOGGER.info("[DEBUG] LLM result type='%s', has content=%s, has tool_name=%s",
-                               result.get("type"), "content" in result, "tool_name" in result)
-
-                    if result["type"] == "tool_call":
-                        LOGGER.info("[DEBUG] Entering tool_call branch")
-                        # Only tool call, no text
-                        if output_keys_spec:
-                            # New behavior: use explicit output_keys
-                            if function_call_key:
-                                state[f"{function_call_key}.name"] = result["tool_name"]
-                                # Store full args dict (for tool_call node dynamic execution)
-                                state[f"{function_call_key}.args"] = result["tool_args"] if isinstance(result["tool_args"], dict) else {}
-                                if isinstance(result["tool_args"], dict):
-                                    for arg_name, arg_value in result["tool_args"].items():
-                                        state[f"{function_call_key}.args.{arg_name}"] = arg_value
-                                        LOGGER.debug("[sea] Stored %s.args.%s = %s", function_call_key, arg_name, arg_value)
-                            # Set conditional_next flags
-                            state["tool_called"] = True
-                            state["has_speak_content"] = False
-                        else:
-                            # Legacy behavior: use predefined keys
-                            state["tool_called"] = True
-                            state["tool_name"] = result["tool_name"]
-                            state["tool_args"] = result["tool_args"]
-                            state["has_speak_content"] = False
-                            # Expand tool_args for legacy args_input (tool_arg_*)
-                            if isinstance(result["tool_args"], dict):
-                                for key, value in result["tool_args"].items():
-                                    state[f"tool_arg_{key}"] = value
-                                    LOGGER.debug("[sea] Expanded tool_arg_%s = %s", key, value)
-
-                        # Format as JSON for logging
-                        text = json.dumps({
-                            "tool": result["tool_name"],
-                            "args": result["tool_args"]
-                        }, ensure_ascii=False)
-                        LOGGER.info("[sea] Tool call detected: %s", text)
-
-                    elif result["type"] == "both":
-                        LOGGER.info("[DEBUG] Entering 'both' branch (text + tool call)")
-                        # Both text and tool call
-                        if output_keys_spec:
-                            # New behavior: use explicit output_keys
-                            if text_key:
-                                state[text_key] = result["content"]
-                                LOGGER.debug("[sea] Stored %s = (text, length=%d)", text_key, len(result["content"]))
-                            if function_call_key:
-                                state[f"{function_call_key}.name"] = result["tool_name"]
-                                # Store full args dict (for tool_call node dynamic execution)
-                                state[f"{function_call_key}.args"] = result["tool_args"] if isinstance(result["tool_args"], dict) else {}
-                                if isinstance(result["tool_args"], dict):
-                                    for arg_name, arg_value in result["tool_args"].items():
-                                        state[f"{function_call_key}.args.{arg_name}"] = arg_value
-                                        LOGGER.debug("[sea] Stored %s.args.%s = %s", function_call_key, arg_name, arg_value)
-                            # Set conditional_next flags
-                            state["tool_called"] = True
-                            state["has_speak_content"] = bool(text_key)
-                        else:
-                            # Legacy behavior: use predefined keys
-                            state["tool_called"] = True
-                            state["tool_name"] = result["tool_name"]
-                            state["tool_args"] = result["tool_args"]
-                            state["has_speak_content"] = True
-                            state["speak_content"] = result["content"]
-                            # Expand tool_args for legacy args_input (tool_arg_*)
-                            if isinstance(result["tool_args"], dict):
-                                for key, value in result["tool_args"].items():
-                                    state[f"tool_arg_{key}"] = value
-                                    LOGGER.debug("[sea] Expanded tool_arg_%s = %s", key, value)
-
-                        text = result["content"]
-                        LOGGER.info("[sea] Both text and tool call detected: tool=%s, text_length=%d",
-                                    result["tool_name"], len(text))
-
-                    else:
-                        LOGGER.info("[DEBUG] Entering 'else' branch (normal text response)")
-                        # Normal text response (no tool call)
-                        state["tool_called"] = False
-                        
-                        if output_keys_spec and text_key:
-                            # New behavior: store in explicit text_key
-                            state[text_key] = result["content"]
-                            LOGGER.info("[sea][llm] Stored state['%s'] = %s", text_key, result["content"])
-                            state["has_speak_content"] = True
-                        else:
-                            # Legacy behavior: no specific text storage (just in "last")
-                            state["has_speak_content"] = True
-
-                        text = result["content"]
-
-                    self._dump_llm_io(playbook.name, getattr(node_def, "id", ""), persona, messages, text)
-                else:
-                    LOGGER.info("[DEBUG] Entering normal mode (no tools)")
-                    # Normal mode (no tools)
-                    state["tool_called"] = False
-
-                    # Check speak flag for streaming output
-                    speak_flag = getattr(node_def, "speak", None)
-                    streaming_enabled = _is_llm_streaming_enabled()
-                    LOGGER.info("[DEBUG] Streaming check: speak_flag=%s, response_schema=%s, streaming_enabled=%s, event_callback=%s",
-                               speak_flag, response_schema is None, streaming_enabled, event_callback is not None)
-                    use_streaming = (
-                        speak_flag is True
-                        and response_schema is None
-                        and streaming_enabled
-                        and event_callback is not None
-                    )
-
-                    if use_streaming:
-                        LOGGER.info("[DEBUG] Using streaming generation (speak=true)")
-                        # Streaming mode: yield chunks to UI (with retry for empty response)
-                        max_stream_retries = 3
-                        text = ""
-                        for stream_attempt in range(max_stream_retries):
-                            text_chunks = []
-                            for chunk in llm_client.generate_stream(
-                                messages,
-                                tools=[],
-                                temperature=self._default_temperature(persona),
-                                **self._get_cache_kwargs(),
-                            ):
-                                # Thinking chunks are dicts, text chunks are strings
-                                if isinstance(chunk, dict) and chunk.get("type") == "thinking":
-                                    event_callback({
-                                        "type": "streaming_thinking",
-                                        "content": chunk["content"],
-                                        "persona_id": getattr(persona, "persona_id", None),
-                                        "node_id": getattr(node_def, "id", "llm"),
-                                    })
-                                    continue
-                                text_chunks.append(chunk)
-                                # Send each text chunk to UI
-                                event_callback({
-                                    "type": "streaming_chunk",
-                                    "content": chunk,
-                                    "persona_id": getattr(persona, "persona_id", None),
-                                    "node_id": getattr(node_def, "id", "llm"),
-                                })
-                            text = "".join(text_chunks)
-
-                            # Check for empty response
-                            if text.strip():
-                                break  # Got valid response
-
-                            # Empty response - discard usage and retry
-                            discarded_usage = llm_client.consume_usage()
-                            LOGGER.warning(
-                                "[sea][llm] Empty streaming response (attempt %d/%d). "
-                                "Discarding usage (in=%d, out=%d) and retrying...",
-                                stream_attempt + 1, max_stream_retries,
-                                discarded_usage.input_tokens if discarded_usage else 0,
-                                discarded_usage.output_tokens if discarded_usage else 0,
-                            )
-                        else:
-                            # All retries exhausted
-                            LOGGER.error(
-                                "[sea][llm] Empty streaming response after %d attempts. "
-                                "Proceeding with empty response.",
-                                max_stream_retries
-                            )
-
-                        # Record usage
-                        usage = llm_client.consume_usage()
-                        LOGGER.info("[DEBUG] consume_usage returned: %s", usage)
-                        llm_usage_metadata: Dict[str, Any] | None = None
-                        if usage:
-                            get_usage_tracker().record_usage(
-                                model_id=usage.model,
-                                input_tokens=usage.input_tokens,
-                                output_tokens=usage.output_tokens,
-                                cached_tokens=usage.cached_tokens,
-                                cache_write_tokens=usage.cache_write_tokens,
-                                cache_ttl=usage.cache_ttl,
-                                persona_id=getattr(persona, "persona_id", None),
-                                building_id=building_id,
-                                node_type="llm_stream",
-                                playbook_name=playbook.name,
-                                category="persona_speak",
-                            )
-                            LOGGER.info("[DEBUG] Usage recorded: model=%s in=%d out=%d cached=%d cache_write=%d", usage.model, usage.input_tokens, usage.output_tokens, usage.cached_tokens, usage.cache_write_tokens)
-                            # Build llm_usage metadata for message
-                            from saiverse.model_configs import calculate_cost, get_model_display_name
-                            cost = calculate_cost(usage.model, usage.input_tokens, usage.output_tokens, usage.cached_tokens, usage.cache_write_tokens, cache_ttl=usage.cache_ttl)
-                            llm_usage_metadata = {
-                                "model": usage.model,
-                                "model_display_name": get_model_display_name(usage.model),
-                                "input_tokens": usage.input_tokens,
-                                "output_tokens": usage.output_tokens,
-                                "cached_tokens": usage.cached_tokens,
-                                "cache_write_tokens": usage.cache_write_tokens,
-                                "cost_usd": cost,
-                            }
-                            # Accumulate into pulse total
-                            self._accumulate_usage(state, usage.model, usage.input_tokens, usage.output_tokens, cost, usage.cached_tokens, usage.cache_write_tokens)
-                        else:
-                            LOGGER.warning("[DEBUG] No usage data from LLM client")
-
-                        # Consume reasoning (thinking) from LLM — store as metadata, not in content
-                        reasoning_entries = llm_client.consume_reasoning()
-                        reasoning_text = "\n\n".join(
-                            e.get("text", "") for e in reasoning_entries if e.get("text")
-                        ) if reasoning_entries else ""
-
-                        # Resolve metadata_key for speak (e.g., media attachments from tool execution)
-                        _speak_metadata_key = getattr(node_def, "metadata_key", None)
-                        _speak_base_metadata = state.get(_speak_metadata_key) if _speak_metadata_key else None
-
-                        # Send completion event with reasoning and metadata
-                        completion_event: Dict[str, Any] = {
-                            "type": "streaming_complete",
-                            "persona_id": getattr(persona, "persona_id", None),
-                            "node_id": getattr(node_def, "id", "llm"),
-                        }
-                        if reasoning_text:
-                            completion_event["reasoning"] = reasoning_text
-                        if _speak_base_metadata and isinstance(_speak_base_metadata, dict):
-                            completion_event["metadata"] = _speak_base_metadata
-                        event_callback(completion_event)
-
-                        # Record to Building history with usage metadata (include pulse total)
-                        pulse_id = state.get("pulse_id")
-                        msg_metadata: Dict[str, Any] = {}
-                        # Merge base metadata first (e.g., media from tool execution)
-                        if _speak_base_metadata and isinstance(_speak_base_metadata, dict):
-                            msg_metadata.update(_speak_base_metadata)
-                        if llm_usage_metadata:
-                            msg_metadata["llm_usage"] = llm_usage_metadata
-                        if reasoning_text:
-                            msg_metadata["reasoning"] = reasoning_text
-                        _at_stream = state.get("_activity_trace")
-                        if _at_stream:
-                            msg_metadata["activity_trace"] = list(_at_stream)
-                        accumulator = state.get("pulse_usage_accumulator")
-                        if accumulator:
-                            msg_metadata["llm_usage_total"] = dict(accumulator)
-                        eff_bid = self._effective_building_id(persona, building_id)
-                        self._emit_say(persona, eff_bid, text, pulse_id=pulse_id, metadata=msg_metadata if msg_metadata else None)
-                    else:
-                        # Non-streaming mode
-                        text = llm_client.generate(
-                            messages,
-                            tools=[],
-                            temperature=self._default_temperature(persona),
-                            response_schema=response_schema,
-                            **self._get_cache_kwargs(),
-                        )
-
-                        # Record usage
-                        usage = llm_client.consume_usage()
-                        llm_usage_metadata: Dict[str, Any] | None = None
-                        if usage:
-                            get_usage_tracker().record_usage(
-                                model_id=usage.model,
-                                input_tokens=usage.input_tokens,
-                                output_tokens=usage.output_tokens,
-                                cached_tokens=usage.cached_tokens,
-                                cache_write_tokens=usage.cache_write_tokens,
-                                cache_ttl=usage.cache_ttl,
-                                persona_id=getattr(persona, "persona_id", None),
-                                building_id=building_id,
-                                node_type="llm",
-                                playbook_name=playbook.name,
-                                category="persona_speak",
-                            )
-                            # Build llm_usage metadata for message
-                            from saiverse.model_configs import calculate_cost, get_model_display_name
-                            cost = calculate_cost(usage.model, usage.input_tokens, usage.output_tokens, usage.cached_tokens, usage.cache_write_tokens, cache_ttl=usage.cache_ttl)
-                            llm_usage_metadata = {
-                                "model": usage.model,
-                                "model_display_name": get_model_display_name(usage.model),
-                                "input_tokens": usage.input_tokens,
-                                "output_tokens": usage.output_tokens,
-                                "cached_tokens": usage.cached_tokens,
-                                "cache_write_tokens": usage.cache_write_tokens,
-                                "cost_usd": cost,
-                            }
-                            # Accumulate into pulse total
-                            self._accumulate_usage(state, usage.model, usage.input_tokens, usage.output_tokens, cost, usage.cached_tokens, usage.cache_write_tokens)
-
-                        # Consume reasoning (thinking) from LLM — store as metadata
-                        reasoning_entries = llm_client.consume_reasoning()
-                        reasoning_text = "\n\n".join(
-                            e.get("text", "") for e in reasoning_entries if e.get("text")
-                        ) if reasoning_entries else ""
-
-                        # If speak=true but streaming disabled, send complete text and record to Building history
-                        LOGGER.info("[DEBUG] speak_flag=%s, event_callback=%s, text_len=%d",
-                                   speak_flag, event_callback is not None, len(text) if text else 0)
-                        if speak_flag is True:
-                            pulse_id = state.get("pulse_id")
-                            # Resolve metadata_key for speak (e.g., media attachments from tool execution)
-                            _speak_metadata_key2 = getattr(node_def, "metadata_key", None)
-                            _speak_base_metadata2 = state.get(_speak_metadata_key2) if _speak_metadata_key2 else None
-                            msg_metadata: Dict[str, Any] = {}
-                            # Merge base metadata first (e.g., media from tool execution)
-                            if _speak_base_metadata2 and isinstance(_speak_base_metadata2, dict):
-                                msg_metadata.update(_speak_base_metadata2)
-                            if llm_usage_metadata:
-                                msg_metadata["llm_usage"] = llm_usage_metadata
-                            if reasoning_text:
-                                msg_metadata["reasoning"] = reasoning_text
-                            _at_speak = state.get("_activity_trace")
-                            if _at_speak:
-                                msg_metadata["activity_trace"] = list(_at_speak)
-                            accumulator = state.get("pulse_usage_accumulator")
-                            if accumulator:
-                                msg_metadata["llm_usage_total"] = dict(accumulator)
-                            eff_bid = self._effective_building_id(persona, building_id)
-                            self._emit_say(persona, eff_bid, text, pulse_id=pulse_id, metadata=msg_metadata if msg_metadata else None)
-                            if event_callback is not None:
-                                LOGGER.info("[DEBUG] Sending 'say' event with content: %s", text[:100] if text else "(empty)")
-                                say_event: Dict[str, Any] = {
-                                    "type": "say",
-                                    "content": text,
-                                    "persona_id": getattr(persona, "persona_id", None),
-                                }
-                                if reasoning_text:
-                                    say_event["reasoning"] = reasoning_text
-                                if _at_speak:
-                                    say_event["activity_trace"] = list(_at_speak)
-                                if msg_metadata:
-                                    say_event["metadata"] = msg_metadata
-                                event_callback(say_event)
-
-                        # Store remaining reasoning for say/speak node (non-speak path)
-                        if reasoning_text:
-                            state["_reasoning_text"] = reasoning_text
-
-                    self._dump_llm_io(playbook.name, getattr(node_def, "id", ""), persona, messages, text)
-                    schema_consumed = self._process_structured_output(node_def, text, state)
-
-                    # Set has_speak_content based on schema_consumed
-                    # If structured output was consumed, we need to set this flag
-                    # Otherwise, it's already set in the tool handling code above
-                    if schema_consumed:
-                        # Structured output means we have usable data, set flag to True
-                        # This allows conditional_next to proceed correctly
-                        state["has_speak_content"] = True
-
-                    # If output_key is specified but no response_schema, store the raw text
-                    if not schema_consumed:
-                        output_key = getattr(node_def, "output_key", None)
-                        if output_key:
-                            state[output_key] = text
-                            LOGGER.info("[sea][llm] Stored plain text to state['%s'] = %s", output_key, text)
-
-                    # Process output_keys even in normal mode (no tools)
-                    output_keys_spec = getattr(node_def, "output_keys", None)
-                    if output_keys_spec:
-                        for mapping in output_keys_spec:
-                            if "text" in mapping:
-                                text_key = mapping["text"]
-                                state[text_key] = text
-                                LOGGER.info("[sea][llm] (normal mode) Stored state['%s'] = %s", text_key, text)
-                                state["has_speak_content"] = True
-                                break
-            except LLMError:
-                # Propagate LLM errors to the caller for proper handling
-                raise
-            except Exception as exc:
-                LOGGER.error("SEA LangGraph LLM failed: %s: %s", type(exc).__name__, exc)
-                # Convert to LLMError so it propagates to the frontend
-                raise LLMError(
-                    f"LLM node failed: {type(exc).__name__}: {exc}",
-                    original_error=exc,
-                ) from exc
-            state["last"] = text
-            # Structured output may return a dict; serialise to JSON string
-            # so that subsequent LLM calls receive valid message content.
-            _msg_content = json.dumps(text, ensure_ascii=False) if isinstance(text, dict) else text
-            state["messages"] = messages + [{"role": "assistant", "content": _msg_content}]
-
-            # Track intermediate messages for profile-based nodes in the same playbook
-            if "_intermediate_msgs" in state:
-                _im = list(state.get("_intermediate_msgs", []))
-                if prompt:
-                    _im.append({"role": "user", "content": prompt})
-                _im.append({"role": "assistant", "content": _msg_content})
-                state["_intermediate_msgs"] = _im
-
-            # Trace: log prompt→response (truncation handled by log_sea_trace)
-            _prompt_str = prompt or "(no prompt)"
-            if schema_consumed:
-                _output_key = getattr(node_def, "output_key", None) or node_id
-                _out_val = state.get(_output_key, text)
-                if isinstance(_out_val, dict):
-                    import json as _json
-                    _resp_str = _json.dumps(_out_val, ensure_ascii=False, default=str)
-                else:
-                    _resp_str = str(_out_val)
-                log_sea_trace(playbook.name, node_id, "LLM", f"prompt=\"{_prompt_str}\" → {_resp_str}")
-            else:
-                _resp_str = str(text) if text else "(empty)"
-                log_sea_trace(playbook.name, node_id, "LLM", f"prompt=\"{_prompt_str}\" → \"{_resp_str}\"")
-
-            # Handle memorize option - save prompt and response to SAIMemory
-            memorize_config = getattr(node_def, "memorize", None)
-            LOGGER.debug("[_lg_llm_node] node=%s memorize_config=%s type=%s schema_consumed=%s", 
-                       getattr(node_def, "id", "?"), memorize_config, type(memorize_config), schema_consumed)
-            if memorize_config:
-                pulse_id = state.get("pulse_id")
-                # Parse memorize config - can be True or {"tags": [...]}
-                if isinstance(memorize_config, dict):
-                    memorize_tags = memorize_config.get("tags", [])
-                else:
-                    memorize_tags = []
-                
-                # Save prompt (user role) - use the pre-expanded prompt variable
-                _memorize_ok = True
-                if prompt:
-                    if not self._store_memory(
-                        persona,
-                        prompt,
-                        role="user",
-                        tags=list(memorize_tags),
-                        pulse_id=pulse_id,
-                    ):
-                        _memorize_ok = False
-                    else:
-                        LOGGER.debug("[sea][llm] Memorized prompt (user): %s", prompt)
-
-                # Save response (assistant role)
-                if text and text != "(error in llm node)":
-                    # If structured output was consumed, format as JSON string for memory
-                    content_to_save = text
-                    if schema_consumed and isinstance(text, dict):
-                        content_to_save = json.dumps(text, ensure_ascii=False, indent=2)
-                        LOGGER.debug("[sea][llm] Structured output formatted as JSON for memory")
-
-                    if not self._store_memory(
-                        persona,
-                        content_to_save,
-                        role="assistant",
-                        tags=list(memorize_tags),
-                        pulse_id=pulse_id,
-                    ):
-                        _memorize_ok = False
-                    else:
-                        LOGGER.debug("[sea][llm] Memorized response (assistant): %s", str(content_to_save))
-
-                if not _memorize_ok and event_callback:
-                    event_callback({"type": "warning", "content": "記憶の保存に失敗しました。会話内容が記録されていない可能性があります。", "warning_code": "memorize_failed", "display": "toast"})
-
-                # Activity trace: record LLM memorize
-                if not playbook.name.startswith(("meta_", "sub_")):
-                    pb_display = playbook.display_name or playbook.name
-                    node_label = getattr(node_def, "label", None) or node_id
-                    _at = state.get("_activity_trace")
-                    if isinstance(_at, list):
-                        _at.append({"action": "memorize", "name": node_label, "playbook": pb_display})
-                    if event_callback:
-                        event_callback({
-                            "type": "activity", "action": "memorize", "name": node_label,
-                            "playbook": pb_display, "status": "completed",
-                            "persona_id": getattr(persona, "persona_id", None),
-                            "persona_name": getattr(persona, "persona_name", None),
-                        })
-
-            # Debug: log speak_content at end of LLM node
-            speak_content = state.get("speak_content", "")
-            LOGGER.info("[DEBUG] LLM node end: state['speak_content'] = '%s'", speak_content)
-
-            # Note: output_mapping in node definition handles state variable assignment
-            # No special handling needed here anymore
-            return state
-
-        return node
+        # Backward-compat: environments running stale runtime_llm bytecode may still
+        # reference module-global _format/_is_llm_streaming_enabled symbols.
+        runtime_llm_module._format = _format
+        runtime_llm_module._is_llm_streaming_enabled = _is_llm_streaming_enabled
+        return lg_llm_node_impl(self, node_def, persona, building_id, playbook, event_callback)
 
     def _default_temperature(self, persona: Any) -> Optional[float]:
         try:
@@ -1205,7 +397,7 @@ class SEARuntime:
 
         # If structured output is needed, check if the selected model supports it
         if needs_structured_output:
-            from saiverse.model_configs import supports_structured_output, get_agentic_model, get_context_length, get_model_provider
+            from saiverse.model_configs import get_agentic_model, get_context_length, get_model_provider, supports_structured_output
             if not supports_structured_output(base_model):
                 # Model doesn't support structured output, switch to agentic model
                 agentic_model = get_agentic_model()
@@ -1232,7 +424,7 @@ class SEARuntime:
 
     def _build_tools_spec(self, tool_names: List[str], llm_client: Any) -> List[Any]:
         """Build tools spec for LLM based on available tool names and llm_client type."""
-        from tools import OPENAI_TOOLS_SPEC, GEMINI_TOOLS_SPEC
+        from tools import GEMINI_TOOLS_SPEC, OPENAI_TOOLS_SPEC
 
         LOGGER.info("[sea] _build_tools_spec called with tool_names: %s", tool_names)
 
@@ -1321,8 +513,8 @@ class SEARuntime:
 
     def _add_playbook_enum(self, schema: Dict[str, Any], available_playbooks_json: str) -> Dict[str, Any]:
         """Dynamically add enum constraint to playbook field in response_schema."""
-        import json
         import copy
+        import json
 
         try:
             # Parse available_playbooks JSON
@@ -1929,7 +1121,8 @@ class SEARuntime:
             LOGGER.debug("[memorize] action_template=%s", action_template)
             LOGGER.debug("[memorize] available variables containing 'finalize': %s", 
                         {k: v for k, v in variables.items() if 'finalize' in str(k).lower()})
-            memo_text = _format(action_template, variables)
+            from .runtime_utils import _format as runtime_format
+            memo_text = runtime_format(action_template, variables)
             LOGGER.debug("[memorize] memo_text=%s", memo_text)
             role = getattr(node_def, "role", "assistant") or "assistant"
             tags = getattr(node_def, "tags", None)
@@ -2090,7 +1283,8 @@ class SEARuntime:
                 "input": state.get("inputs", {}).get("input", ""),
                 "last": state.get("last", ""),
             })
-            sub_input = _format(template, variables)
+            from .runtime_utils import _format as runtime_format
+            sub_input = runtime_format(template, variables)
             eff_bid = self._effective_building_id(persona, building_id)
 
             # Determine execution mode
@@ -2196,7 +1390,8 @@ class SEARuntime:
 
         # Simple template expansion
         try:
-            result = _format(value_template, state)
+            from .runtime_utils import _format as runtime_format
+            result = runtime_format(value_template, state)
             if result == value_template and "{" in value_template:
                 # Template was not expanded - log for debugging
                 LOGGER.debug("[sea][set] Template not expanded. Keys in state: %s", list(state.keys()))
@@ -2367,7 +1562,8 @@ class SEARuntime:
 
             node_id = getattr(node_def, "id", "stelis_start")
             label_raw = getattr(node_def, "label", None) or "Stelis Session"
-            label = _format(label_raw, state)
+            from .runtime_utils import _format as runtime_format
+            label = runtime_format(label_raw, state)
 
             # Send status event
             if event_callback:
@@ -3177,10 +2373,10 @@ class SEARuntime:
     ) -> None:
         """Generate Chronicle entries from all unprocessed messages."""
         from llm_clients.factory import get_llm_client
-        from saiverse.model_configs import find_model_config
         from sai_memory.arasuji import init_arasuji_tables
-        from sai_memory.arasuji.generator import ArasujiGenerator, DEFAULT_BATCH_SIZE
+        from sai_memory.arasuji.generator import DEFAULT_BATCH_SIZE, ArasujiGenerator
         from sai_memory.memory.storage import get_messages_paginated
+        from saiverse.model_configs import find_model_config
 
         # Get LLM client using MEMORY_WEAVE_MODEL
         model_name = os.getenv("MEMORY_WEAVE_MODEL", "gemini-2.5-flash-lite-preview-09-2025")
@@ -3231,450 +2427,16 @@ class SEARuntime:
     # ---------------- context preparation -----------------
 
     def _prepare_context(self, persona: Any, building_id: str, user_input: Optional[str], requirements: Optional[Any] = None, pulse_id: Optional[str] = None, warnings: Optional[List[Dict[str, Any]]] = None, preview_only: bool = False) -> List[Dict[str, Any]]:
-        from sea.playbook_models import ContextRequirements
-
-        # Use provided requirements or default to full context
-        reqs = requirements if requirements else ContextRequirements()
-
-        messages: List[Dict[str, Any]] = []
-
-        # ---- system prompt ----
-        if reqs.system_prompt:
-            system_sections: List[str] = []
-
-            # 1. Common prompt (world setting, framework explanation)
-            common_prompt_template = getattr(persona, "common_prompt", None)
-            LOGGER.debug("common_prompt_template is %s (type=%s)", common_prompt_template, type(common_prompt_template))
-            if common_prompt_template:
-                try:
-                    # Get building info for variable expansion
-                    building_obj = getattr(persona, "buildings", {}).get(building_id)
-                    building_name = building_obj.name if building_obj else building_id
-                    city_name = getattr(persona, "current_city_id", "unknown_city")
-
-                    # Expand variables in common prompt using safe replace (avoid conflict with JSON examples)
-                    common_text = common_prompt_template
-                    replacements = {
-                        "{current_persona_name}": getattr(persona, "persona_name", "Unknown"),
-                        "{current_persona_id}": getattr(persona, "persona_id", "unknown_id"),
-                        "{current_building_name}": building_name,
-                        "{current_city_name}": city_name,
-                        "{current_persona_system_instruction}": getattr(persona, "persona_system_instruction", ""),
-                        "{current_building_system_instruction}": getattr(building_obj, "system_instruction", "") if building_obj else "",
-                        "{linked_user_name}": getattr(persona, "linked_user_name", "the user"),
-                    }
-                    for placeholder, value in replacements.items():
-                        common_text = common_text.replace(placeholder, value)
-                    system_sections.append(common_text.strip())
-                except Exception as exc:
-                    LOGGER.error("Failed to format common prompt: %s", exc, exc_info=True)
-
-            # 2. "## あなたについて" section
-            persona_section_parts: List[str] = []
-            persona_sys = getattr(persona, "persona_system_instruction", "") or ""
-            if persona_sys:
-                persona_section_parts.append(persona_sys.strip())
-
-            # persona inventory
-            if reqs.inventory:
-                try:
-                    inv_builder = getattr(persona, "_inventory_summary_lines", None)
-                    inv_lines: List[str] = inv_builder() if callable(inv_builder) else []
-                except Exception:
-                    inv_lines = []
-                if inv_lines:
-                    persona_section_parts.append("### インベントリ\n" + "\n".join(inv_lines))
-
-            if persona_section_parts:
-                system_sections.append("## あなたについて\n" + "\n\n".join(persona_section_parts))
-
-            # 3. "## {building_name}" section (current location)
-            try:
-                building_obj = getattr(persona, "buildings", {}).get(building_id)
-                if building_obj:
-                    building_section_parts: List[str] = []
-
-                    # Building system instruction
-                    # NOTE: Datetime variables ({current_time}, etc.) are no longer expanded here.
-                    # Time information is now provided via Realtime Context at the end of messages
-                    # to improve LLM context caching efficiency.
-                    # Use base_system_instruction (without items) to avoid duplication
-                    # with the building_items block below.
-                    building_sys = getattr(building_obj, "base_system_instruction", None) or getattr(building_obj, "system_instruction", None)
-                    if building_sys:
-                        building_section_parts.append(str(building_sys).strip())
-
-                    # Building items
-                    if reqs.building_items:
-                        try:
-                            items_by_building = getattr(self.manager, "items_by_building", {}) or {}
-                            item_registry = getattr(self.manager, "item_registry", {}) or {}
-                            b_items = items_by_building.get(building_id, [])
-                            lines = []
-                            for iid in b_items:
-                                data = item_registry.get(iid, {})
-                                raw_name = data.get("name", "") or ""
-                                name = raw_name.strip() if raw_name.strip() else "(名前なし)"
-                                desc = (data.get("description") or "").strip() or "(説明なし)"
-                                lines.append(f"- [{iid}] {name}: {desc}")
-                            if lines:
-                                building_section_parts.append("### 建物内のアイテム\n" + "\n".join(lines))
-                        except Exception:
-                            LOGGER.warning("Failed to collect building items for %s", building_id, exc_info=True)
-
-                    if building_section_parts:
-                        building_name = getattr(building_obj, "name", building_id)
-                        system_sections.append(f"## {building_name} (ID: {building_id})\n" + "\n\n".join(building_section_parts))
-            except Exception:
-                LOGGER.warning("Failed to build building section for system prompt", exc_info=True)
-
-            # 4. "## 利用可能な能力" section (available playbooks)
-            if reqs.available_playbooks:
-                try:
-                    from tools import TOOL_REGISTRY
-                    list_playbooks_func = TOOL_REGISTRY.get("list_available_playbooks")
-                    if list_playbooks_func:
-                        # Get available playbooks JSON (tool returns string; accept old tuple form)
-                        playbooks_raw = list_playbooks_func(
-                            persona_id=getattr(persona, "persona_id", None),
-                            building_id=building_id
-                        )
-                        playbooks_json = playbooks_raw[0] if isinstance(playbooks_raw, tuple) else playbooks_raw
-                        if playbooks_json:
-                            import json
-                            playbooks_list = json.loads(playbooks_json)
-                            if playbooks_list:
-                                playbooks_formatted = json.dumps(playbooks_list, ensure_ascii=False, indent=2)
-                                system_sections.append(f"## 利用可能な能力\n以下のPlaybookを実行できます：\n```json\n{playbooks_formatted}\n```")
-                except Exception as exc:
-                    LOGGER.debug("Failed to add available playbooks section: %s", exc)
-
-            # 5. "## 現在の状況" section (working memory)
-            if reqs.working_memory:
-                try:
-                    sai_mem = getattr(persona, "sai_memory", None)
-                    if sai_mem and sai_mem.is_ready():
-                        wm_data = sai_mem.load_working_memory()
-                        if wm_data:
-                            import json as json_mod
-                            wm_text = json_mod.dumps(wm_data, ensure_ascii=False, indent=2)
-                            system_sections.append(f"## 現在の状況\n```json\n{wm_text}\n```")
-                            LOGGER.debug("[sea][prepare-context] Added working_memory section")
-                except Exception as exc:
-                    LOGGER.debug("Failed to add working_memory section: %s", exc)
-
-            # NOTE: Spatial context (Unity) has been moved to Realtime Context
-            # to improve LLM context caching efficiency.
-
-            system_text = "\n\n---\n\n".join([s for s in system_sections if s])
-            if system_text:
-                messages.append({"role": "system", "content": system_text})
-
-        # ---- Memory Weave context (Chronicle + Memopedia) ----
-        # Inserted between system prompt and visual context
-        _mw_persona_enabled = self._is_memory_weave_context_enabled(persona) if reqs.memory_weave else False
-        LOGGER.info("[sea][prepare-context] memory_weave=%s, persona_enabled=%s", reqs.memory_weave, _mw_persona_enabled)
-        if reqs.memory_weave and _mw_persona_enabled:
-            try:
-                from builtin_data.tools.get_memory_weave_context import get_memory_weave_context
-                from tools.context import persona_context
-                persona_id = getattr(persona, "persona_id", None)
-                
-                # Get persona_dir from sai_memory adapter (same pattern as working_memory)
-                sai_mem = getattr(persona, "sai_memory", None)
-                persona_dir_path = getattr(sai_mem, "persona_dir", None) if sai_mem else None
-                persona_dir = str(persona_dir_path) if persona_dir_path else None
-                
-                LOGGER.info("[sea][prepare-context] Calling get_memory_weave_context for persona=%s dir=%s", persona_id, persona_dir)
-                with persona_context(persona_id, persona_dir, self.manager):
-                    mw_messages = get_memory_weave_context(persona_id=persona_id, persona_dir=persona_dir)
-                LOGGER.info("[sea][prepare-context] get_memory_weave_context returned %d messages", len(mw_messages))
-                if mw_messages:
-                    messages.extend(mw_messages)
-                    LOGGER.debug("[sea][prepare-context] Added %d Memory Weave context messages", len(mw_messages))
-            except Exception as exc:
-                LOGGER.exception("[sea][prepare-context] Failed to get Memory Weave context: %s", exc)
-
-        # ---- visual context (Building / Persona images) ----
-        # Inserted right after system prompt but before conversation history
-        if reqs.visual_context:
-            try:
-                from builtin_data.tools.get_visual_context import get_visual_context
-                from tools.context import persona_context, get_active_manager
-                persona_id = getattr(persona, "persona_id", None)
-                persona_dir = getattr(persona, "persona_dir", None)
-                with persona_context(persona_id, persona_dir, self.manager):
-                    visual_messages = get_visual_context(building_id=building_id)
-                if visual_messages:
-                    messages.extend(visual_messages)
-                    LOGGER.debug("[sea][prepare-context] Added %d visual context messages", len(visual_messages))
-            except Exception as exc:
-                LOGGER.debug("[sea][prepare-context] Failed to get visual context: %s", exc)
-
-        # ---- history ----
-        history_depth = reqs.history_depth
-        if history_depth not in [0, "none"]:
-            history_mgr = getattr(persona, "history_manager", None)
-            if history_mgr:
-                try:
-                    # Determine which tags to include
-                    required_tags = ["conversation"]
-                    if reqs.include_internal:
-                        required_tags.append("internal")
-
-                    # Parse history_depth format
-                    # - "full": use max_history_messages (message count) or context_length (character limit)
-                    # - "Nmessages" (e.g., "10messages"): message count limit
-                    # - integer or numeric string: character limit
-                    use_message_count = False
-                    limit_value = 2000  # fallback
-                    used_anchor = False
-                    recent = []
-
-                    if history_depth == "full":
-                        metabolism_enabled = getattr(self.manager, "metabolism_enabled", False) if self.manager else False
-
-                        if metabolism_enabled and not preview_only:
-                            # Persistent anchor resolution with 3-level fallback
-                            anchor_id, resolution = self._resolve_metabolism_anchor(persona)
-
-                            if anchor_id:
-                                # Case 1 or 2: valid anchor found
-                                recent_from_anchor = history_mgr.get_history_from_anchor(
-                                    anchor_id, required_tags=required_tags, pulse_id=pulse_id,
-                                )
-                                if recent_from_anchor:
-                                    recent = recent_from_anchor
-                                    used_anchor = True
-                                    history_mgr.metabolism_anchor_message_id = anchor_id
-                                    LOGGER.debug(
-                                        "[sea][prepare-context] Anchor-based retrieval (%s): %d messages from anchor %s",
-                                        resolution, len(recent), anchor_id,
-                                    )
-                                    # Persist anchor for current model (touch updated_at)
-                                    persona_model = getattr(persona, "model", None)
-                                    if persona_model:
-                                        self._update_anchor_for_model(persona, persona_model, anchor_id)
-                            else:
-                                # Case 3: no valid anchor — minimal load + Chronicle generation
-                                memory_weave_enabled = os.getenv("ENABLE_MEMORY_WEAVE_CONTEXT", "").lower() in ("true", "1")
-                                if memory_weave_enabled and self._is_chronicle_enabled_for_persona(persona):
-                                    try:
-                                        LOGGER.info("[metabolism] Triggering Chronicle generation on anchor expiry")
-                                        self._generate_chronicle(persona)
-                                    except Exception as exc:
-                                        LOGGER.warning("[metabolism] Chronicle generation on anchor expiry failed: %s", exc)
-
-                                # Load minimal history (low watermark)
-                                low_wm = self._get_low_watermark(persona)
-                                limit_value = low_wm if low_wm and low_wm > 0 else 20
-                                use_message_count = True
-                                LOGGER.debug(
-                                    "[sea][prepare-context] Minimal load (no valid anchor): %d messages",
-                                    limit_value,
-                                )
-
-                        elif metabolism_enabled and preview_only:
-                            # Preview mode: use anchor for retrieval but don't persist or generate Chronicle
-                            anchor_id, resolution = self._resolve_metabolism_anchor(persona)
-                            if anchor_id:
-                                recent_from_anchor = history_mgr.get_history_from_anchor(
-                                    anchor_id, required_tags=required_tags, pulse_id=pulse_id,
-                                )
-                                if recent_from_anchor:
-                                    recent = recent_from_anchor
-                                    used_anchor = True
-                            if not used_anchor:
-                                low_wm = self._get_low_watermark(persona)
-                                limit_value = low_wm if low_wm and low_wm > 0 else 20
-                                use_message_count = True
-
-                        if not used_anchor and not metabolism_enabled:
-                            # Metabolism disabled — traditional count/char retrieval
-                            max_hist_msgs = getattr(self.manager, "max_history_messages_override", None) if self.manager else None
-                            if max_hist_msgs is None:
-                                from saiverse.model_configs import get_default_max_history_messages
-                                persona_model = getattr(persona, "model", None)
-                                if persona_model:
-                                    max_hist_msgs = get_default_max_history_messages(persona_model)
-                            if max_hist_msgs is not None:
-                                limit_value = max_hist_msgs
-                                use_message_count = True
-                                LOGGER.debug("[sea][prepare-context] Using max_history_messages=%d", max_hist_msgs)
-                            else:
-                                limit_value = getattr(persona, "context_length", 2000)
-
-                    elif isinstance(history_depth, str) and history_depth.endswith("messages"):
-                        # Message count mode: "10messages", "20messages", etc.
-                        try:
-                            limit_value = int(history_depth[:-8])  # Remove "messages" suffix
-                            use_message_count = True
-                        except ValueError:
-                            limit_value = 10  # fallback for message count
-                            use_message_count = True
-                    else:
-                        try:
-                            limit_value = int(history_depth)
-                        except (ValueError, TypeError):
-                            limit_value = 2000  # fallback
-
-                    # Fetch history if not already retrieved via anchor
-                    if not used_anchor:
-                        LOGGER.debug("[sea][prepare-context] Fetching history: limit=%d, mode=%s, pulse_id=%s, balanced=%s, tags=%s",
-                                    limit_value, "messages" if use_message_count else "chars", pulse_id, reqs.history_balanced, required_tags)
-
-                        if use_message_count:
-                            # Message count mode - balanced not supported yet
-                            recent = history_mgr.get_recent_history_by_count(
-                                limit_value,
-                                required_tags=required_tags,
-                                pulse_id=pulse_id,
-                            )
-                        elif reqs.history_balanced:
-                            # Get conversation partners for balanced retrieval
-                            participant_ids = ["user"]
-                            occupants = self.manager.occupants.get(building_id, [])
-                            persona_id = getattr(persona, "persona_id", None)
-                            for oid in occupants:
-                                if oid != persona_id:
-                                    participant_ids.append(oid)
-                            LOGGER.debug("[sea][prepare-context] Balancing across: %s", participant_ids)
-                            recent = history_mgr.get_recent_history_balanced(
-                                limit_value,
-                                participant_ids,
-                                required_tags=required_tags,
-                                pulse_id=pulse_id,
-                            )
-                        else:
-                            # Filter by required tags or current pulse_id
-                            recent = history_mgr.get_recent_history(
-                                limit_value,
-                                required_tags=required_tags,
-                                pulse_id=pulse_id,
-                            )
-
-                        # Set metabolism anchor on first count-based retrieval and persist (skip in preview)
-                        metabolism_enabled_for_anchor = getattr(self.manager, "metabolism_enabled", False) if self.manager else False
-                        if metabolism_enabled_for_anchor and recent and not preview_only:
-                            oldest_id = recent[0].get("id")
-                            if oldest_id:
-                                history_mgr.metabolism_anchor_message_id = oldest_id
-                                persona_model = getattr(persona, "model", None)
-                                if persona_model:
-                                    self._update_anchor_for_model(persona, persona_model, oldest_id)
-                                LOGGER.debug("[sea][prepare-context] Set metabolism anchor to %s (persisted)", oldest_id)
-
-                    LOGGER.debug("[sea][prepare-context] Got %d history messages", len(recent))
-                    # Enrich messages with attachment context
-                    enriched_recent = self._enrich_history_with_attachments(recent)
-                    messages.extend(enriched_recent)
-                except Exception as exc:
-                    LOGGER.exception("[sea][prepare-context] Failed to get history: %s", exc)
-
-        # ---- Realtime Context ----
-        # Time-sensitive info placed just BEFORE the last user message to improve LLM caching.
-        # This ensures LLM responds to user input, not the realtime context.
-        if reqs.realtime_context:
-            try:
-                realtime_msg = self._build_realtime_context(persona, building_id, messages)
-                if realtime_msg:
-                    # Find the last user message and insert realtime context before it
-                    last_user_idx = None
-                    for i in range(len(messages) - 1, -1, -1):
-                        if messages[i].get("role") == "user" and not messages[i].get("metadata", {}).get("__realtime_context__"):
-                            last_user_idx = i
-                            break
-
-                    if last_user_idx is not None:
-                        # Insert before last user message
-                        messages.insert(last_user_idx, realtime_msg)
-                        LOGGER.debug("[sea][prepare-context] Added realtime context before last user message (idx=%d)", last_user_idx)
-                    else:
-                        # No user message found, append at end
-                        messages.append(realtime_msg)
-                        LOGGER.debug("[sea][prepare-context] Added realtime context at end (no user message found)")
-            except Exception as exc:
-                LOGGER.debug("[sea][prepare-context] Failed to build realtime context: %s", exc)
-
-        # ---- Token budget check ----
-        try:
-            from saiverse.token_estimator import estimate_messages_tokens
-            from saiverse.model_configs import get_context_length, get_model_provider
-
-            persona_model = getattr(persona, "model", None)
-            if persona_model:
-                provider = get_model_provider(persona_model)
-                context_length = get_context_length(persona_model)
-                estimated_tokens = estimate_messages_tokens(messages, provider)
-                LOGGER.debug(
-                    "[sea][prepare-context] Token budget: estimated=%d, limit=%d (model=%s)",
-                    estimated_tokens, context_length, persona_model,
-                )
-
-                if estimated_tokens > context_length:
-                    # Over budget: trim history messages from oldest until within budget
-                    # Find indices of history messages (not system, not visual context, not realtime)
-                    history_indices = []
-                    for i, msg in enumerate(messages):
-                        meta = msg.get("metadata") or {}
-                        if (
-                            msg.get("role") != "system"
-                            and not meta.get("__visual_context__")
-                            and not meta.get("__realtime_context__")
-                            and not meta.get("__memory_weave_context__")
-                        ):
-                            history_indices.append(i)
-
-                    original_count = len(history_indices)
-                    # Remove oldest history messages until under budget
-                    while history_indices and estimated_tokens > context_length:
-                        remove_idx = history_indices.pop(0)
-                        removed_msg = messages[remove_idx]
-                        removed_tokens = estimate_messages_tokens([removed_msg], provider)
-                        estimated_tokens -= removed_tokens
-                        messages[remove_idx] = None  # mark for removal
-
-                    # Clean up None entries
-                    messages = [m for m in messages if m is not None]
-                    remaining_count = len(history_indices)
-
-                    warning_msg = {
-                        "type": "warning",
-                        "warning_code": "context_auto_trimmed",
-                        "content": (
-                            f"コンテキスト超過のため、履歴を直近{original_count}件→{remaining_count}件に"
-                            f"自動削減しました（推定: {estimated_tokens:,} / {context_length:,}トークン）。"
-                            f"ChatOptionsでメッセージ数上限を下げてください。"
-                        ),
-                    }
-                    LOGGER.warning(
-                        "[sea][prepare-context] Context auto-trimmed: %d -> %d messages (est=%d, limit=%d)",
-                        original_count, remaining_count, estimated_tokens, context_length,
-                    )
-                    if warnings is not None:
-                        warnings.append(warning_msg)
-
-                elif estimated_tokens > context_length * 0.85:
-                    # Approaching limit: warn but continue
-                    warning_msg = {
-                        "type": "warning",
-                        "warning_code": "context_approaching_limit",
-                        "content": (
-                            f"コンテキスト使用量がモデルの上限に近づいています"
-                            f"（推定: {estimated_tokens:,} / {context_length:,}トークン）。"
-                            f"ChatOptionsでメッセージ数上限を下げることを検討してください。"
-                        ),
-                    }
-                    LOGGER.warning(
-                        "[sea][prepare-context] Context approaching limit: est=%d, limit=%d (%.0f%%)",
-                        estimated_tokens, context_length, estimated_tokens / context_length * 100,
-                    )
-                    if warnings is not None:
-                        warnings.append(warning_msg)
-        except Exception as exc:
-            LOGGER.debug("[sea][prepare-context] Token budget check failed: %s", exc)
-
-        return messages
+        return prepare_context_impl(
+            self,
+            persona,
+            building_id,
+            user_input,
+            requirements=requirements,
+            pulse_id=pulse_id,
+            warnings=warnings,
+            preview_only=preview_only,
+        )
 
     # ---- Context Preview (read-only, no side effects) ----
 
@@ -3683,174 +2445,9 @@ class SEARuntime:
         persona: Any,
         building_id: str,
         user_input: str,
-        meta_playbook: Optional[str] = None,
-        image_count: int = 0,
-        document_count: int = 0,
+        playbook_name: Optional[str] = None,
     ) -> Dict[str, Any]:
-        """Build the context that would be sent to the LLM, without executing anything.
-
-        Returns a dict with messages, token estimates, cost estimates, and model info.
-        Does NOT record the user message to history or call any LLM.
-        """
-        from saiverse.token_estimator import estimate_messages_tokens, estimate_image_tokens
-        from saiverse.model_configs import (
-            get_model_provider, get_context_length,
-            get_model_display_name, get_model_pricing, calculate_cost,
-        )
-
-        # Select playbook (same logic as run_meta_user)
-        if meta_playbook:
-            playbook = self._load_playbook_for(meta_playbook, persona, building_id)
-            if playbook is None:
-                playbook = self._choose_playbook(kind="user", persona=persona, building_id=building_id)
-        else:
-            playbook = self._choose_playbook(kind="user", persona=persona, building_id=building_id)
-
-        # Build context messages (without recording user message to history)
-        # Use "conversation" profile requirements to match what sub_speak actually sees,
-        # not the meta-playbook's own context_requirements (which may lack memory_weave etc.)
-        from sea.playbook_models import CONTEXT_PROFILES
-        preview_requirements = CONTEXT_PROFILES["conversation"]["requirements"]
-        context_warnings: List[Dict[str, Any]] = []
-        messages = self._prepare_context(
-            persona, building_id, user_input=None,
-            requirements=preview_requirements,
-            warnings=context_warnings,
-            preview_only=True,
-        )
-
-        # Append the user message manually (in real flow it comes from history)
-        if user_input:
-            messages.append({"role": "user", "content": user_input})
-
-        # Classify each message into a section
-        persona_model = getattr(persona, "model", None) or "gemini-2.5-flash-lite-preview-09-2025"
-        provider = get_model_provider(persona_model)
-
-        section_order = [
-            "system_prompt", "memory_weave_chronicle", "memory_weave_memopedia",
-            "memory_weave", "visual_context",
-            "history", "realtime_context", "user_message",
-        ]
-        section_labels = {
-            "system_prompt": "System Prompt",
-            "memory_weave_chronicle": "Memory Weave — Chronicle",
-            "memory_weave_memopedia": "Memory Weave — Memopedia",
-            "memory_weave": "Memory Weave",
-            "visual_context": "Visual Context",
-            "history": "Conversation History",
-            "realtime_context": "Realtime Context",
-            "user_message": "Your Message",
-            "attachments": "Attachments",
-        }
-        section_tokens: Dict[str, int] = {s: 0 for s in section_order}
-        section_tokens["attachments"] = 0
-        section_msg_counts: Dict[str, int] = {s: 0 for s in section_order}
-        section_msg_counts["attachments"] = 0
-
-        annotated_messages: List[Dict[str, Any]] = []
-        for i, msg in enumerate(messages):
-            meta = msg.get("metadata") or {}
-            # Determine section
-            if msg.get("role") == "system":
-                section = "system_prompt"
-            elif meta.get("__memory_weave_context__"):
-                mw_type = meta.get("__memory_weave_type__", "")
-                if mw_type == "chronicle":
-                    section = "memory_weave_chronicle"
-                elif mw_type == "memopedia":
-                    section = "memory_weave_memopedia"
-                else:
-                    section = "memory_weave"
-            elif meta.get("__visual_context__"):
-                section = "visual_context"
-            elif meta.get("__realtime_context__"):
-                section = "realtime_context"
-            elif i == len(messages) - 1 and msg.get("role") == "user" and user_input and msg.get("content") == user_input:
-                section = "user_message"
-            else:
-                section = "history"
-
-            msg_tokens = estimate_messages_tokens([msg], provider)
-            section_tokens[section] += msg_tokens
-            section_msg_counts[section] += 1
-
-            annotated_messages.append({
-                "role": msg.get("role", "unknown"),
-                "content": msg.get("content", ""),
-                "section": section,
-                "tokens": msg_tokens,
-            })
-
-        # Add estimated attachment tokens
-        attachment_tokens = 0
-        if image_count > 0:
-            attachment_tokens += image_count * estimate_image_tokens(provider)
-        if document_count > 0:
-            # Rough estimate: ~500 tokens per document (varies widely)
-            attachment_tokens += document_count * 500
-        section_tokens["attachments"] = attachment_tokens
-
-        total_input_tokens = sum(section_tokens.values())
-        context_length = get_context_length(persona_model)
-        pricing = get_model_pricing(persona_model)
-
-        # Cost range: best case (all cached) to worst case (all cache-write)
-        cache_kwargs = self._get_cache_kwargs()
-        cache_enabled = cache_kwargs.get("enable_cache", False)
-        cache_ttl = cache_kwargs.get("cache_ttl", "5m")
-
-        # Determine cache type (explicit for Anthropic, implicit for Gemini, etc.)
-        from saiverse.model_configs import get_cache_config
-        cache_config = get_cache_config(persona_model)
-        cache_type = cache_config.get("type", "implicit")
-
-        if cache_enabled and pricing and pricing.get("cached_input_per_1m_tokens") is not None:
-            # Best case: everything is a cache hit
-            cost_best = calculate_cost(
-                persona_model, total_input_tokens, 0,
-                cached_tokens=total_input_tokens, cache_write_tokens=0,
-            )
-            # Worst case: everything is a cache write
-            cost_worst = calculate_cost(
-                persona_model, total_input_tokens, 0,
-                cached_tokens=0, cache_write_tokens=total_input_tokens,
-                cache_ttl=cache_ttl,
-            )
-        else:
-            # No cache: single estimate
-            cost_best = calculate_cost(persona_model, total_input_tokens, 0)
-            cost_worst = cost_best
-
-        # Build sections summary
-        all_sections = section_order + ["attachments"]
-        sections_summary = []
-        for s in all_sections:
-            if section_tokens.get(s, 0) > 0 or section_msg_counts.get(s, 0) > 0:
-                sections_summary.append({
-                    "name": s,
-                    "label": section_labels.get(s, s),
-                    "tokens": section_tokens.get(s, 0),
-                    "message_count": section_msg_counts.get(s, 0),
-                })
-
-        return {
-            "persona_id": getattr(persona, "persona_id", "unknown"),
-            "persona_name": getattr(persona, "persona_name", "Unknown"),
-            "model": persona_model,
-            "model_display_name": get_model_display_name(persona_model),
-            "provider": provider,
-            "context_length": context_length,
-            "sections": sections_summary,
-            "total_input_tokens": total_input_tokens,
-            "estimated_cost_best_usd": round(cost_best, 6),
-            "estimated_cost_worst_usd": round(cost_worst, 6),
-            "cache_enabled": cache_enabled,
-            "cache_ttl": cache_ttl if cache_enabled else None,
-            "cache_type": cache_type if cache_enabled else None,
-            "pricing": pricing or {},
-            "messages": annotated_messages,
-        }
+        return preview_context_impl(self, persona, building_id, user_input, playbook_name=playbook_name)
 
     def _enrich_history_with_attachments(self, messages: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
         """Enrich history messages with attachment context.
